- Mistakenly messed up the Old Task type definition so this was reverted to the correct definition - Update supabase table definitions again for typescript to include is current for current task - Update all get tasks queries to include updated at column for tasks and is current for current task queries where needed. - set all the fields for adding and editting that were being hardcoded or using placeholder now that everythjing is in place. - Add check that task is created by user as a safety precaution for adding, editting and deleting

+++ app/manage/list/components/task_queries/DeleteTaskQuery.tsx
@@ -1,10 +1,11 @@
 import supabaseClient from "@/lib/supabaseClient";
 
-const DeleteTaskQuery = async (taskId: string) => {
+const DeleteTaskQuery = async (taskId: string, userId: string) => {
   const { error: supabaseError } = await supabaseClient
     .from("tasks")
     .delete()
-    .eq("id", taskId);
+    .eq("id", taskId)
+    .eq("created_by", userId);
 
   if (supabaseError) {
     throw new Error(supabaseError.message);
